Fixed typo introduced in commit #1108040

+++ shaders/fragment.glsl
@@ -6,7 +6,7 @@
 
 #if defined(GL_ES)
 #if defined(GL_FRAGMENT_PRECISION_HIGH)
-precision mediump float;	//ES2.0 supporting highp
+precision highp float;	//ES2.0 supporting highp
 #else
 precision mediump float;	//Default
 #endif
